requirements.txt
alembic==1.4.1
anyio==3.2.1
apache-airflow==2.1.2
apache-airflow-providers-ftp==2.0.0
apache-airflow-providers-imap==2.0.0
apache-airflow-providers-sqlite==2.0.0
apispec==3.3.2
argcomplete==1.12.3
attrs==20.3.0
Babel==2.9.1
blinker==1.4
cachetools==4.2.2
cattrs==1.5.0
certifi==2020.12.5
cffi==1.14.5
charset-normalizer==2.0.4
click==7.1.2
clickclick==20.10.2
cliff==3.8.0
cloudpickle==1.6.0
cmaes==0.8.2
cmd2==2.1.2
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
croniter==1.0.15
cryptography==3.4.7
databricks-cli==0.14.3
defusedxml==0.7.1
dill==0.3.2
dnspython==1.16.0
docker==5.0.0
docutils==0.16
email-validator==1.1.3
entrypoints==0.3
fastavro==1.4.4
feast==0.12.0
Flask==1.1.4
Flask-AppBuilder==3.3.1
Flask-Babel==1.0.0
Flask-Caching==1.10.1
Flask-JWT-Extended==3.25.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.14.3
gitdb==4.0.7
GitPython==3.1.18
google-api-core==1.31.1
google-auth==1.34.0
googleapis-common-protos==1.52.0
graphviz==0.16
grpcio==1.39.0
gunicorn==20.1.0
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
idna==2.10
importlib-resources==1.5.0
inflection==0.5.1
iso8601==0.1.14
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.3
joblib==1.0.1
jsonschema==3.2.0
lazy-object-proxy==1.4.3
lockfile==0.12.2
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
marshmallow==3.12.2
marshmallow-enum==1.5.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.23.1
mlflow==1.19.0
mmh3==3.0.0
numpy==1.21.0
openapi-schema-validator==0.1.5
openapi-spec-validator==0.3.1
optuna==2.9.1
packaging==21.0
pandas==1.3.0
pandavro==1.5.2
pbr==5.6.0
pendulum==2.1.2
prettytable==2.1.0
prison==0.1.3
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
psutil==5.8.0
pyarrow==5.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydantic==1.8.2
Pygments==2.9.0
PyJWT==1.7.1
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.18.0
python-daemon==2.3.0
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python3-openid==3.2.0
pytz==2021.1
pytzdata==2020.1
PyYAML==5.4.1
querystring-parser==1.2.4
requests==2.26.0
rfc3986==1.5.0
rich==10.5.0
rsa==4.7.2
scikit-learn==0.24.2
scipy==1.7.1
setproctitle==1.2.2
six==1.16.0
sklearn==0.0
smmap==4.0.0
sniffio==1.2.0
SQLAlchemy==1.3.24
SQLAlchemy-JSONField==1.0.0
SQLAlchemy-Utils==0.37.8
sqlparse==0.4.1
stevedore==3.3.0
swagger-ui-bundle==0.0.8
tabulate==0.8.9
tenacity==8.0.1
termcolor==1.1.0
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
tqdm==4.62.0
typer==0.3.2
typing-extensions==3.10.0.0
unicodecsv==0.14.1
urllib3==1.26.6
wcwidth==0.2.5
websocket-client==1.1.0
Werkzeug==1.0.1
WTForms==2.3.3
